+++ resources/DisallowedRanges.php
@@ -12,10 +12,6 @@ final class DisallowedRanges
             return true;
         }
 
-        if ($codePoint >= 2155 && $codePoint <= 2207) {
-            return true;
-        }
-
         if ($codePoint >= 3676 && $codePoint <= 3712) {
             return true;
         }
@@ -32,15 +28,11 @@ final class DisallowedRanges
             return true;
         }
 
-        if ($codePoint >= 6849 && $codePoint <= 6911) {
+        if ($codePoint >= 6863 && $codePoint <= 6911) {
             return true;
         }
 
-        if ($codePoint >= 11859 && $codePoint <= 11903) {
-            return true;
-        }
-
-        if ($codePoint >= 42955 && $codePoint <= 42996) {
+        if ($codePoint >= 11870 && $codePoint <= 11903) {
             return true;
         }
 
@@ -76,11 +68,11 @@ final class DisallowedRanges
             return true;
         }
 
-        if ($codePoint >= 66928 && $codePoint <= 67071) {
+        if ($codePoint >= 67005 && $codePoint <= 67071) {
             return true;
         }
 
-        if ($codePoint >= 67432 && $codePoint <= 67583) {
+        if ($codePoint >= 67515 && $codePoint <= 67583) {
             return true;
         }
 
@@ -112,7 +104,7 @@ final class DisallowedRanges
             return true;
         }
 
-        if ($codePoint >= 69466 && $codePoint <= 69551) {
+        if ($codePoint >= 69514 && $codePoint <= 69551) {
             return true;
         }
 
@@ -136,7 +128,7 @@ final class DisallowedRanges
             return true;
         }
 
-        if ($codePoint >= 71488 && $codePoint <= 71679) {
+        if ($codePoint >= 71495 && $codePoint <= 71679) {
             return true;
         }
 
@@ -168,7 +160,7 @@ final class DisallowedRanges
             return true;
         }
 
-        if ($codePoint >= 75076 && $codePoint <= 77823) {
+        if ($codePoint >= 75076 && $codePoint <= 77711) {
             return true;
         }
 
@@ -177,10 +169,6 @@ final class DisallowedRanges
         }
 
         if ($codePoint >= 83527 && $codePoint <= 92159) {
-            return true;
-        }
-
-        if ($codePoint >= 92784 && $codePoint <= 92879) {
             return true;
         }
 
@@ -200,11 +188,11 @@ final class DisallowedRanges
             return true;
         }
 
-        if ($codePoint >= 101641 && $codePoint <= 110591) {
+        if ($codePoint >= 101641 && $codePoint <= 110575) {
             return true;
         }
 
-        if ($codePoint >= 110879 && $codePoint <= 110927) {
+        if ($codePoint >= 110883 && $codePoint <= 110927) {
             return true;
         }
 
@@ -212,7 +200,11 @@ final class DisallowedRanges
             return true;
         }
 
-        if ($codePoint >= 113828 && $codePoint <= 118783) {
+        if ($codePoint >= 113828 && $codePoint <= 118527) {
+            return true;
+        }
+
+        if ($codePoint >= 118724 && $codePoint <= 118783) {
             return true;
         }
 
@@ -224,7 +216,11 @@ final class DisallowedRanges
             return true;
         }
 
-        if ($codePoint >= 121520 && $codePoint <= 122879) {
+        if ($codePoint >= 121520 && $codePoint <= 122623) {
+            return true;
+        }
+
+        if ($codePoint >= 122655 && $codePoint <= 122879) {
             return true;
         }
 
@@ -232,11 +228,11 @@ final class DisallowedRanges
             return true;
         }
 
-        if ($codePoint >= 123216 && $codePoint <= 123583) {
+        if ($codePoint >= 123216 && $codePoint <= 123535) {
             return true;
         }
 
-        if ($codePoint >= 123648 && $codePoint <= 124927) {
+        if ($codePoint >= 123648 && $codePoint <= 124895) {
             return true;
         }
 
@@ -276,10 +272,6 @@ final class DisallowedRanges
             return true;
         }
 
-        if ($codePoint >= 129751 && $codePoint <= 129791) {
-            return true;
-        }
-
         if ($codePoint >= 129995 && $codePoint <= 130031) {
             return true;
         }
@@ -288,7 +280,7 @@ final class DisallowedRanges
             return true;
         }
 
-        if ($codePoint >= 173790 && $codePoint <= 173823) {
+        if ($codePoint >= 173792 && $codePoint <= 173823) {
             return true;
         }